Split `constants.ts` into `app.ts` and `game.ts`
