+++ src/phases/attempt-capture-phase.ts
@@ -1,4 +1,4 @@
-import { PLAYER_PARTY_MAX_SIZE } from "#app/constants/constants";
+import { PLAYER_PARTY_MAX_SIZE } from "#app/constants/game";
 import type { SubstituteTag } from "#app/data/battler-tags/substitute-tag";
 import {
   doPokeballBounceAnim,
