+++ src/plugins/api/account-api.ts
@@ -4,7 +4,7 @@ import type {
   AccountLoginResponse,
   AccountRegisterRequest,
 } from "#app/@types/AccountApi";
-import { SESSION_ID_COOKIE } from "#app/constants/constants";
+import { SESSION_ID_COOKIE } from "#app/constants/app";
 import { ApiBase } from "#app/plugins/api/api-base";
 import { removeCookie, setCookie } from "#app/utils";
