+++ src/ui/handlers/login-form-ui-handler.ts
@@ -1,4 +1,4 @@
-import { APP_ABBREVIATION, SAVE_FILE_EXTENSION, SAVES_ZIP_PREFIX } from "#app/constants/constants";
+import { APP_ABBREVIATION, SAVE_FILE_EXTENSION, SAVES_ZIP_PREFIX } from "#app/constants/app";
 import { globalScene } from "#app/global-scene";
 import { api } from "#app/plugins/api/api";
 import { GAME_HEIGHT, GAME_WIDTH } from "#app/constants/ui";
