+++ src/ui/handlers/run-info-ui-handler.ts
@@ -1,5 +1,5 @@
 import type { SessionSaveData } from "#app/@types/SessionData";
-import { PLAYER_PARTY_MAX_SIZE } from "#app/constants/constants";
+import { PLAYER_PARTY_MAX_SIZE } from "#app/constants/game";
 import { getBiomeName } from "#app/data/biome-utils";
 import { getNatureName, getNatureStatMultiplier } from "#app/data/nature";
 import { getPokeballAtlasKey } from "#app/data/pokeball";
