+++ src/ui/settings/move-touch-controls-handler.ts
@@ -1,4 +1,4 @@
-import { LS_PREFIX } from "#app/constants/constants";
+import { LS_PREFIX } from "#app/constants/app";
 import { eventBus } from "#app/event-bus";
 import { globalScene } from "#app/global-scene";
 import type TouchControl from "#app/touch-controls";
