+++ test/data/splash-messages.test.ts
@@ -1,6 +1,6 @@
 import { getSplashMessages } from "#app/data/splash-messages";
 import { describe, expect, it, vi, afterEach, beforeEach } from "vitest";
-import * as Constants from "#app/constants/constants";
+import * as Constants from "#app/constants/app";
 
 describe("Data - Splash Messages", () => {
   it("should contain at least 15 splash messages", () => {
