+++ test/test-utils/testFileInitialization.ts
@@ -1,5 +1,5 @@
 import { initLoggedInUser } from "#app/account";
-import { SESSION_ID_COOKIE } from "#app/constants/constants";
+import { SESSION_ID_COOKIE } from "#app/constants/app";
 import { initEggMoves } from "#app/data/egg-moves";
 import { allMoves } from "#app/data/data-lists";
 import { initAbilities } from "#app/data/init/init-abilities";
